+++ app/core/models/__init__.py
@@ -31,3 +31,5 @@ from .ai_commerce import *
 
 from .ai_agents import *
 from .ai_product import *
+
+from app.core.models.market import *